pyproject.toml
[build-system]
requires = ["maturin>=1.8,<2.0"]
build-backend = "maturin"

[project]
name = "yoco"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dynamic = ["version"]
dependencies = [
    "pillow>=10.4.0",
    "tqdm>=4.67.1",
]
[project.optional-dependencies]
tests = [
    "pytest",
]
[tool.maturin]
python-source = "python"
features = ["pyo3/extension-module"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
